big refactoring

+++ src/main/java/com/goapi/goapi/repo/database/DatabaseRepo.java
@@ -1,16 +1,13 @@
 package com.goapi.goapi.repo.database;
 
-import com.goapi.goapi.domain.database.Database;
-import com.goapi.goapi.domain.user.User;
+import com.goapi.goapi.domain.model.database.Database;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import javax.xml.crypto.Data;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface DatabaseRepo extends JpaRepository<Database, Integer> {
 
-    List<Database> findAllByOwnerId(Long ownerId);
+    List<Database> findAllByOwnerId(Integer ownerId);
 }
